+++ src/AsyncMqttClient_Generic_Debug.h
@@ -9,7 +9,7 @@
   
   Built by Khoi Hoang https://github.com/khoih-prog/AsyncMqttClient_Generic
  
-  Version: 1.3.0
+  Version: 1.4.0
   
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -19,6 +19,7 @@
   1.2.0    K Hoang     15/03/2022 Add support to STM32 using LAN8742A (without TLS/SSL)
   1.2.1    K Hoang     16/03/2022 Add support to STM32 using LAN8720 (without TLS/SSL)
   1.3.0    K Hoang     16/03/2022 Add support to Portenta_H7 using built-in Ethernet or Murata WiFi (without TLS/SSL)
+  1.4.0    K Hoang     17/03/2022 Add support to Teensy 4.1 using QNEthernet Library
  *****************************************************************************************************************************/
 
 #pragma once
@@ -51,6 +52,14 @@
 
 #define AMQTT_PRINT           DBG_PORT_AMQTT.print
 #define AMQTT_PRINTLN         DBG_PORT_AMQTT.println
+#define AMQTT_PRINTF          DBG_PORT_AMQTT.printf
+
+/////////////////////////////////////////////////////////
+
+#define AMQTT_DEBUG(...) 				if(_ASYNC_MQTT_LOGLEVEL_>3) { AMQTT_PRINTF(__VA_ARGS__); }
+#define AMQTT_SSL_DEBUG(...) 		if(_ASYNC_MQTT_LOGLEVEL_>3) { AMQTT_PRINTF(__VA_ARGS__); }
+
+#define AMQTT_ASSERT( a )       do{ if(!(a)){AMQTT_PRINTF("ASSERT: %s %u \n", __FILE__, __LINE__);}}while(0)
 
 /////////////////////////////////////////////////////////